Extract Store::{get_job, list_jobs} etc into impl module.
Following f0a0d20dbd2b24eb0dcde82169d22a271261190b and previous commits,
continues extracting methods off of Store and into separate files with
logical groups of methods, shrinking the size of store.rs. The methods
here are all read-only methods.
